donate some tests to BlockExecutionSpec
